[REF] website, web_editor: confine 'NULL' "hack" into website
Commit [1] introduced a "hack" to transform `'NULL'` values into `null`
values in some maps as normally-defined `null` values are automatically
removed (replaced by a fallback). This hack did not need to be general
and in web_editor but can rather be confined to website only for the
user palette maps (especially since the fix made by the parent of this
commit here).

[1]: https://github.com/odoo/odoo/commit/dfd6c60c830ef4c625474e90df54a2a103039d0a

closes odoo/odoo#76033

Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>