[FIX] crm: prevent division by zero in lead probability computation
If for some reason we cannot compute lead probability, we should simply
continue to the next value without causing a division by zero.

For example, this can happen if all stages are team specific, as there is
a current limitation regarding the first stage (used to know how many lost
and won there is) that requires to have no team assigned to it. This is a
side effect of the commit https://github.com/odoo/odoo/commit/67c49bf3af08d480780f0406461b305c5e50559f introduced in V14, and we should add
a test to ensure we do not crash in that case.

closes odoo/odoo#76266

X-original-commit: 7f3a4680f12f6db5c9a229abaea1acf1d1a0fd96
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>
Signed-off-by: Alex Tuyls <alt-odoo@users.noreply.github.com>