[FIX] calendar: display start and end of events without datetimerange widget
Before this commit, the datetime range widget was used to set these values in for calendar form view.
Unfortunately, some glitches appeared in mobile view. This commit, revert this change because fixing the mobile view for this specific case was not trivial.
Datetime range widget works in mobile for regular form view (not launched from the calendar view).

closes odoo/odoo#76086

closes odoo/odoo#76977

Taskid: 2635704
Signed-off-by: Kevin Baptiste <kba@odoo.com>