[FIX] web: pivot: can flip axis while reloading the data
Before this commit, if the user clicked on flip axis while the
data was being reloaded (e.g. just after applying a filter), the
reload was cancelled, and the axis wasn't flipped.

With this commit, we wait for the data to be loaded, and then
flip the axis.

Part-of: odoo/odoo#76146