[FIX] google_recaptcha: Add key in session_info
Before this commit, if there was a Google ReCaptcha public key
registered on the database, it was given to the backend along with
the assets in the `session_info` global key.

Since this key is meant to be used on the frontend to generate recaptcha
tokens, it should be available there as well.

Now, it is also given to the frontend version of the `session_info` key.

closes odoo/odoo#76596

Task: 2639095
X-original-commit: 58ceeeceb11de5f3b393d26eff629d8b790cea4a
Signed-off-by: Julien Mougenot (jum) <jum@odoo.com>