[FIX] account: use current date as fallback to convert currency
Description of the issue/feature this PR addresses: Fixes #76177

Current behavior before PR:
Before this commit, There will be tracback on adding new
Journal Item if there is not date on Journal Entry.

Desired behavior after PR is merged:
Now we use current date to as fallback to convert currency.

closes odoo/odoo#76406

X-original-commit: ad2c5b06176c60bcd0a71b514530a8b6569cd638
Signed-off-by: oco-odoo <oco-odoo@users.noreply.github.com>