[FIX] website_event_track_live: fix widescreen detection for YouTube tracks
Currently, when there is a YouTube URL configured on the track, we
automatically set the layout of the track page as "widescreen".

This implies that we hide some menus and adapt the layout to display the full
video player.

However, in the XML template of the track view, we only display the video
player if there is a YouTube URL configured AND one or more of those conditions
is respected:
- The track is configured as a "YouTube Replay"
- The track is "coming soon" (< 30 minutes)
- The track is live
- The track is done

This is done to match the two behaviors and get a consistent display.

Task-2643147

closes odoo/odoo#76293

X-original-commit: d342b21aa7890da3e2283020d841690e01be012d
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>