[FIX] web: favorite domain is a string
Since the refactoring of the search model https://github.com/odoo/odoo/commit/e807e3582372d9ff4157bffbf660bd3e1898c795,
the new favorite items were created with a domain with the wrong type:
Array instead of string. Thus it was impossible to share the search
state with a legacy view when a new favorite was created. We now set
correctly the domain key of a favorite.

closes odoo/odoo#76311

Related: odoo/enterprise#20767
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>