[FIX] web: control panel bottom content style
Before this commit, the style of the control panel's bottom section
was a bit broken: the button groups did not flex.
This was because the dom changed a bit.

After this commit, they do.
Also, the structure of control-panel-bottom-left has changed.
It should contain `.btn-group` elements at its root.
The spacing between groups is handled by the control-panel.
(before, those spaces were actual spaces in DOM)

Part-of: odoo/odoo#76342