[FIX] hr_holidays: fix not being able to save added_value_type
The readonly attribute in form views makes the field not be saved upon
change. Since the field was present two times, once for changes and a
second time just for display, and the second one was readonly, the field
would never be updated on the server.

TaskId-2637489

Part-of: odoo/odoo#77311