[FIX] web_tour: Give correct access when reading modules
Tours are automatically disabled when the demo data are enabled on
at least one module. To ensure that, when giving the session_info
object, we read on the `ir.module.module` model to check whether some
of them have demo data.

Before this commit, this check only relied on the user's `is_admin`
flag, which does not give read access on the `ir.module.module` model.

Now, a `sudo` ensures that any admin user can read on the model.

closes odoo/odoo#76857

X-original-commit: 619cee59f52894d37277a3a108cce01ea5646717
Signed-off-by: Rémi Rahir (rar) <rar@odoo.com>