[FIX] website: prevent animations from restarting on a blank page
Commit [1] introduced `_onEditionWasStopped` which besides declaring the
edit mode as disabled, also restarted animations. Since the new editor
was introduced, this is triggered on save and cancel, which is indeed
needed in order to declare that edit mode is disabled, but also
generates a traceback due to the animations now restarting on an already
destroyed wysiwyg. That restarting of animations is unnecessary anyway
since the page will refresh when disabling the edit mode.

[1]: https://github.com/odoo/odoo/commit/ff6d3feba7237fb1b2136fef02ecd57847b8dadb

task-2528814

closes odoo/odoo#76446

X-original-commit: 7b9aacc6582cc83f3c9f149f8f7257bf8084a7e3
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>