[FIX] web: falsy group names in list view
When a group has a falsy value (e.g. if grouped by an integer field having a "0" group), make sure to display its proper name instead of "Undefined".
Off course, boolean fields are still showing "false" name if it is the case.

This issue was introduced with ecb2f32

closes odoo/odoo#75692

Related: odoo/enterprise#20477
Signed-off-by: Mathieu Duckerts-Antoine <Polymorphe57@users.noreply.github.com>