Revert "[FIX] mail,web: Change command palette shortcut"
This reverts commit 46f648c34a974a263fc47d520be75b4d350070c9.

The command+m shortcut on safari is not listenable. Safari prevents to
catch events of some shortcut reserved for OS actions. Command+m is
used to minimize the app. This makes it impossible to open the command
palette on safari. So we decided to use again the command+k (ctrl+k)
to open the command palette.

Part-of: odoo/odoo#76478