[FIX] base: en_GB, week_start: Monday
Monday is commonly accepted as the first day of the week in the UK.

Listed as first option on https://www.unicode.org/cldr/cldr-aux/charts/29/supplemental/territory_information.html#GB with Sunday as an alternative.
All mayor OS (Windows, Mac, iOS) use Monday as the first day of the week when the locale en_GB is used.

closes odoo/odoo#76764

X-original-commit: 394885de715548e3d04d28caf510d4da00e23233
Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>