[FIX] website_crm_partner_assign: hide activity types of other models
Right now when editing the opportunity, all the activity types are listed in
the drop-down, even those which are specific to other models.

After this commit  while editing the opportunity, we now display
only activity types which are either specific for 'crm.lead', or those
which are not linked with any models (common for all).

TaskId-2443894

closes odoo/odoo#77920

X-original-commit: 1bf0ea8d0295b400a17ce197bdde7ba3f682966c
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>