[FIX] hr_holidays: fix wrong context for absent action
Since odoo/odoo#35700 department_id is not a searchable field in the
employee search view as it has been added in the searchpanel. Since
then, the view from department > absence has been broken since it was
not filtering on the department anymore.
This commit fixes that behaviour by defaulting to the department in the
searchpanel instead.

TaskId-2648380

closes odoo/odoo#77152

X-original-commit: 396abfb4578f165d63287c3ce72d7539938b2977
Signed-off-by: Kevin Baptiste <kba@odoo.com>