[FIX] web_editor: properly await all async parts for onRemove/onClone
Even though onRemove and onClone cannot be async themselves (yet?), it
is called via a trigger_up of `call_for_each_child_snippet` so that each
part of the snippet being removed / cloned has their onRemove / onClone
called. Doing that way, some SnippetEditor instances may have to be
created and it can be an async operation... the problem is that our code
was only awaiting the first of those instanciations instead of all of
them.

closes odoo/odoo#76879

X-original-commit: c7e3484f60d1ab55791b84d4a1c69bfcf0a16fd8
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>