[FIX] website: avoid error when no template installed for dynamic snip.
Since [1] if neither website_blog nor website_sale was installed, upon
selecting a dynamic snippet, an error was raised.
It therefore prevented deleting the unusable snippet.

After this commit the error is avoided by coping with the fact that it
is possible that no template is installed.

[1] https://github.com/odoo/odoo/commit/086571123cc04e410d3f0845b3516c130dbec189

Related to #75662

task-2651978

closes odoo/odoo#76828

Signed-off-by: Jérémy Kersten (jke) <jke@openerp.com>